Extract TreeMap functions into utils
